Splits morph::Visual into and ownable and a final class, getting rid of all those horrible OWNED_MODEs
